+++ mvc/src/main/java/com/interface21/context/SimpleApplicationContext.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.handler;
+package com.interface21.context;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
